requirements-server.txt
# LockIn Web Server Requirements

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Limiter==3.5.0
Werkzeug==2.3.7
marshmallow==3.20.1
requests==2.31.0

# Production WSGI server
gunicorn==21.2.0

# PostgreSQL driver (used in production / OpenShift)
psycopg2-binary==2.9.9
